backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==15.0.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.43.0
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
faster-whisper==1.1.1
filelock==3.18.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
mistune==3.1.3
mpmath==1.3.0
numpy==2.3.1
onnxruntime==1.22.0
openai==1.93.3
packaging==24.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sympy==1.14.0
tokenizers==0.21.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
